move function that touch db to the db folder
